0028263: Coding Rules - clean up definition of the class Graphic3d_MaterialAspect
Fixed uninitialized fields in several classes.
Body of trivial methods have been moved to class definition (header file).
Ensure that non-primitive types are returned by reference, when possible.
Removed unused class Prs3d_PlaneSet.

+++ src/OpenGl/OpenGl_Context.cxx
@@ -2877,8 +2877,8 @@ void OpenGl_Context::SetShadingMaterial (const OpenGl_AspectFace* theAspect,
   }
 
   // handling transparency
-  float aTranspFront = (float )aMatFrontSrc.Transparency();
-  float aTranspBack  = (float )aMatBackSrc .Transparency();
+  float aTranspFront = aMatFrontSrc.Transparency();
+  float aTranspBack  = aMatBackSrc .Transparency();
   if (!theHighlight.IsNull()
     && theHighlight->BasicFillAreaAspect().IsNull())
   {
